% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{consensus}
\alias{consensus}
\title{\code{consensus} returns taxa common to both dataset x and dataset y}
\usage{
consensus(x, y)
}
\arguments{
\item{x, y}{Tibbles output from \code{metaxize}}
}
\description{
\code{consensus} returns taxa common to both dataset x and dataset y
}
