% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch_proc.R
\name{load_blast6}
\alias{load_blast6}
\title{\code{load_blast6} imports blast6-formatted VSEARCH output files from a given directory and merges them into a single tibble}
\usage{
load_blast6(x)
}
\arguments{
\item{x}{A file path to directory containing (only) VSEARCH output files}
}
\description{
\code{load_blast6} imports blast6-formatted VSEARCH output files from a given directory and merges them into a single tibble
}
