% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsearch_proc.R
\name{top_hit}
\alias{top_hit}
\title{\code{top_hit} extracts the top ranked subject sequence for each query sequence per sample}
\usage{
top_hit(x)
}
\arguments{
\item{x}{A tibble produced by \code{load_blast6}}
}
\value{
A tibble
}
\description{
\code{top_hit} extracts the top ranked subject sequence for each query sequence per sample
}
